Some minor safe mode improvements
Background crashes no longer count toward entering safe mode.  Safe
mode will only be entered after the app crashes twice in a row while
being launched into the foreground.  Background crashes are not as
visible to the user, who might be confused about the reference to a
crash on the safe mode screen when the app hasn't crashed.

Improve the message displayed to the user on the safe mode screen.  It
now gives the user a clue as to how long safe mode might last.

Log both an error message and a support request when entering safe
mode.  Both of these increase the chances that a developer will notice
and investigate the problem.  (The support request does not contain
the user's e-mail address, since the user has not given permission to
send it.  Its purpose is to provide an easy and visible link to the
telemetry.)

Keep the safe mode screen visible for at least two seconds, even if
telemetry and HockeyApp are already caught up, so that the user has
time to read the screen.  (Seeing a screen flash with the word "crash"
on it but not having time to read the whole message could be
disconcerting for a user.)

Fix https://support.nachocove.com/helpdesk/tickets/786