Fix McMutables race condition that caused a crash
If two different threads called McMutables.Set() at the same time with
the same key, it could result in two database items with the same key,
which would result in a crash the next time that key was accessed.

Fix the code to avoid the race condition, using DB transactions as the
synchronization method.

Add code to detect and correct old databases that have duplicate
McMutables items.

Fix nachocove/qa#1863