Code improvement to CheckedChange callbacks
The Android Switch.CheckedChange event is called when the switch is
set to its initial value as the UI is being configured, but the
callback code was assuming that all calls were the result of the user
changing the value.  This resulted in extra work being done.  Fix the
callbacks to only do their thing when the change came from the user.