Don't crash when dismissed view wants to show alert dialog
A fragment that has been dismissed has a null Activity, which means it
doesn't have an easy way to get a usable Context, which means it can't
easily show an alert dialog.  Check for a null Activity in fragment
callback functions that might be called after the fragment has been
dismissed.

Fix nachocove/qa#1875