better handling of the SyncKit.PendingSingle
NcCommand subclasses have some built-in resolving functions, which rely
on NcCommand.PendingSingle and NcCommand.PendingList. By also using
SyncKit.PendingSingle, things got a bit confused.
When we create the command, set the SyncKit.PendingSingle to null so we
aren’t tempted to use it later.
Then, if we resolved the pending ourselves, set NcCommand.PendingSingle
to null, so that any later error paths don’t try to resolve it as
failed or deferred (which is what caused the crash).

Further, if we overlooked a case, log an error instead of crashing.
resolves nachocove/qa#1893