HTML signatures on iOS
iOS now supports either plain text or HTML signatures for an account.
HTML signatures cannot be edited, only copy/pasted.

Due to time constraints, this task is not complete.  The UI for the
signature page is ugly and needs some work.  The code would benefit
from some refactoring.  Android still needs to be done.  But the iOS
functionality should be there.