Catch exceptions from cancelling a token
The exception could be related to some socket error, because we try to,
in some cases, read and/or close the socket on cancellation, which
might fail if we’ve survived across shutdown.
Also pull up CmdName and CmdNameWithAccount to NcCommand, and adjust
code accordingly.
Might help shed light on nachocove/qa#1925