get the cropped contact picture on iOS
It's smaller than the original, but still plenty large for our use (a few hundred pixels per edge).

Without this, a lot of contact photos appear squished if they weren't already squares to begin with.

An alternative would be to get the orginal image, the crop box, and crop it ourselves, but I can't find any way to get the crop box