Database performance improvements
Remove 83 indexes that (1) are never used, or (2) have a small number
of distinct values such as indexes for boolean columns.  Such indexes
take time to maintain when the database is updated, but have little or
no benefit when running queries.  Removing the [Indexed] attribute
from the code does not automatically delete the index from existing
databases, so a migration was created to delete the indexes manually.

Tweak a number of McEmailMessage queries:
1. References to the McEmailMessageDependency table were removed sice
I couldn't find any uses of that table.
2. likelihood values were adjusted to help the query planner pick the
most appropriate index.