Improve code for PerformFetch shutdown and badge/notification
Several changes to the lifecycle for PerformFetch and to badge count /
message notification code.

Use the iOS background time remaining (rather than a fixed-length
25-second timer) to determine when perform fetch is running out of
time and needs to shut down.

Run the DB queries for badge count / message notification on a
background thread rather than the UI thread.  This prevents a slow
query from causing the app to run past its alotted time.

Update the badge count when the app transitions from foreground to
background, in case the badge count became inaccurate while the app
was in the foreground.