Fix missing 'return' for not-enough-time-left quick sync case
When a quick sync is started with very little background time
remaining, the app doesn't even try to do the quick sync, since there
is a good chace that it won't finish successfully.  But that code was
missing a 'return' statement, which led to the quick sync happening
anyway, but with a null completion handler, which resulted in a crash
when the quick sync was done.

Fix nachocove/qa#1968