Avoid race condition crash in Android foreground/background detector
The callback for a timer is not run as soon as the timer fires, but is
queued up to run as a task in the thread pool.  That means the
callback can run after the timer has been disposed.  Fix
GoToBackgroundTimer in class LifecycleSpy to detect this situation and
to handle it gracefully rather than crash.

Fix nachocove/qa#1975