Clean up stale DB connections
The code that manages database connections was recently revamped.  The
new code is simpler and cleaner.  But it does have the potential to
leak connections over time, which could eventually lead to the app
crashing due to too many open files.

This update adds code to identify and close DB connections that have
been leaked.  The cleanup code does not use a periodic timer (as has
been done in the past).  Instead, the cleanup code runs whenever the
app goes into the background.  The leaks should be slow enough that an
infrequent cleanup like this should be enough.