Move some message list view queries from UI thread to background thread
For some of the message list views on iOS, the database queries that
retrieve the list of messages to display have been moved from the UI
thread to a background thread.  This was done for the unified inbox
and for any view of a single folder (including a single-account
inbox).  These queries are usually fast, but sometimes they can be
slow enough to cause problems if run on the UI thread.  Other message
list views, such as deferred messages or a single thread, still run
their queries on the UI thread.