Overhaul how brain events are scheduled
The scheduling of brain events has been overhauled.  There is very
little change to the work that the brain does.  The changes are in how
the brain decides what to do and when to do it.

Many events that were previously persisted in the database are now
stored only in the brain's in-memory event queue.  This avoids a
number of database transactions that had been happening on the UI
thread or in very tiny background tasks.  There is a chance that some
of these events could be lost when the app is killed, but that is
unlikely because the processing of those events is the brain's top
priority.  And if a few of the events are lost, the effect on the user
is minimal.

The work that the brain was doing as part of its periodic glean and
when new email messages arrived is now incorporated into the brain's
main processing loop.  There is still a timer that wakes up the brain
from time to time so it can look for work to do, but the brain's
work is now done as soon as possible rather than only happening in
these periodic episodes.

The brain now prioritizes its work in a reasonable way.  Scoring and
gleaning of messages is near the top.  Indexing of contacts is at the
bottom.  Indexing of messages is in the middle.  (The details are more
complicated.  See NcBrain.Process().)  The brain will no longer spend
a couple of hours indexing contacts before it starts scoring messages.

The brain now runs all the time on iOS, not just when the app is in
the foreground.  But it restricts what it does when in the background,
and does only the very highest priority work when the app is in quick
sync mode.

Fix nachocove/qa#1984