Open contact detail when tapping on a name in the message detail
In the message detail view on iOS, when the user taps on the From
field or on any of the names in the To or Cc fields, open the
appropriate contact detail view.  If the user taps elsewhere in the
header, toggle between the compact and expanded header views, as was
the case before.