Message view swipe actions are now responsive and don't pause
On both iOS and Android, change the message list view code so that the
Delete, Archive, and Move swipe actions are always responsive.  As
soon as the user taps one of those actions, the UI removes the selected
message(s) from the list and returns to its normal state.  Database
changes no longer happen on the UI thread.  The message list is updated
immediately rather than waiting for the status indicator event from
the back end.

Fix nachocove/qa#2041