Don't crash on startup when user had changed the tabs
NachoTabBarController.RestoreCustomTabBarOrder() depends on code in
NachoTabBarController's constructor.  But RestoreCustomTabBarOrder()
is called indirectly through the base class constructor and therefore
runs before NachoTabBarController's constructor.  This causes a crash
on app startup if the user has changed the tabs at the bottom of the
screen.

To avoid this problem, move all the code from the constructor to
ViewDidLoad().

Fix nachocove/qa#2054