Avoid infinite loop of Sync / FSync commands
Detect a cycle of Sync / FSync commands that are not accomplishing
anything, and break the cycle.  This protects against misbehaving
ActiveSync servers that either send sync status of FolderChange_12
when they shouldn't or don't send folder changes when they should.

It is unknown if anything useful can be done with the account when the
app ignores the sync status FolderChange_12.  It may be that nothing
can be synched.  But at least it won't go into an infinite loop.

Fix nachocove/qa#2057