fix race condition causing blank inbox
Here's what would happen:
- inbox controller would construct and note the current account
- user would switch accounts before viewing inbox
- user would view inbox
- viewDidLoad would run and call Reload using the old account
- viewWillAppear would run and notice the account difference and call SwitchToAccount
- SwitchToAccoutn would blank out the message source (by creating a new one)
- SwitchToAccount would expect viewWillAppear to call Reload
- viewWillAppear doesn't call Reload on first load because we want to call it as early as possible (in viewDidLoad)
- HandleReload would run and see an empty source

The first fix checks the account in viewDidLoad and switches if necessary before the first Reload.  This fixes the race
condition described above.

The second fix, calling SetNeedsReload in SwitchToAccount, makes sure that if a reload is pending while the user switches
accounts, another reload will be queued.  It also has the effect of preemptnig the Reload in viewDidLoad or viewDidAppear
when they notice an account difference, but it's not a problem to call reload a few lines earlier in those cases.