Make sure the message body view isn't permanently hidden
When a message download fails, the UIWebView that shows the message
body might be marked as hidden.  When that same UIWebVie was reused
for a different message, it was not being unhidden, which meant the
message body was blank.  There was no way to get out of this situation
without restarting the app.

Change MessageViewController.LoadView() to make sure the UIWebView
object is not hidden.