Get rid of the brain's periodic timer
Now that the brain keeps processing events until there are no more
events left to process, there is no real need for a timer that wakes
up the brain from time to time.  As long as the brain's status
indicator event listener listens for enough events, the brain will
wake up often enough to never get too far behind.

This change reduces the frequency with which the app is woken up while
in the background on Android, and should therefore reduce the app
battery usage.