Don't allow sending a message to an obviously bogus address
When composing a message, make sure that all the recipients' email
addresses contain a '@' character.  With this change, the user is told
about the bogus address before sending the message, rather than having
the message sit in the outbox forever where the user will never see
it.

For technical reasons, the behavior is a little different on iOS and
Android.  On iOS, the user is told when the bogus address is first
entered.  On Android, the user is told about the problem when the Send
button is tapped.

The same checking is also done when adding attendees to a meeting.

Fix nachocove/qa#2044