Avoid stuck refresh indicator in message list view
When an iOS user pulls down to refresh in a message list view and then
immediately opens a message detail while the refresh is in progress,
the refresh indicator would still be visible, and would not go away,
when the user returned to the message list view.

To avoid an orphaned refresh indicator, always clean up the refresh
indicator when a message list view disappears.

Fix https://support.nachocove.com/helpdesk/tickets/997