Limit description field to four lines when event editor is opened
(All of these changes apply to both iOS and Android.)

When the event editor is opened, show a maximum of four lines of text
for the description field.  The description field is expanded to its
full height, showing the entire text, when the user moves the focus to
that field.  The user will always have the full text available when
changing the description.  Once the description field is expanded, it
is never shrunk again.  Having the description field start out
truncated guarantees that the user can see that there are very useful
fields below the description.

Do a better job of extracting the plain text out of an e-mail message
when creating a meeting from a message.

When the user edits an event description within the app, show the new
value for the description.  Previously, changes to the description
would only be shown if they came from the server.  Changes made in the
app were sent to the server, but were not visible in the app.
Fix nachocove/qa#1727