Use McEmailAddress.Score when ranking contacts/email search results
When searching for an email address, such as message compose
autocomplete or meeting attendees, consider the McEmailAddress.Score
field when ranking the search results.  How well the contact matches
the search string still dominates the rankings; the email address
score is used to distinguish between contacts whose search string
match score is similar.

(The general contact search algorithm, used when searching on the
contacts tab, was not changed and does not use McEmailAddress.Score.
This change only affects when searching for an email address.)