Update build config for new android signing key
Previously we signed the app with a key we controlled, and then uploaded
to google.  Using that method, the key can never be changed without
creating a whole new app in the store.  Now Google allows using one key
to upload to them, and they re-sign with another key that they manage
for distrubution.  So we will be able to change the upload key but keep
the same distrubition key