revert documents migration
The original move was so extensions would have direct access to the main database.
After working with extensions, it seems like they'll function best if they have
pre-processed data sets so very little work needs to be done in the extension itself.
Therefore, we'll move the main data back, and only populated the shared area with
small data sets relevant to the extensions.