imap tweaks after updating mailkit
- the \NoSelect flag is missing sometimes when it used to be there
- But the \NonExistent flag is still there, and it implies \NoSelect
- So always check for \NoSelect OR \NonExistent to tell if we can select a folder

Incorrect check was causing our code to try to examine the [Gmail] folder during a
folder sync command, which would result in an error from the gmail imap server
and stop all imap activity.

Not sure exactly what changed, but I'm guessing it was something to do with MailKit's
internal folder storage, because we get \NoSelect for [Gmail] in one response, and
then get \NonExistent in another.  It's like the older MailKit would merge the two
responses, but the newer one doesn't.